requirements.txt
bcrypt==3.1.0
pycrypto==2.6.1
pyDes==2.0.1

